Implemented Headless Chrome to enable server-side automation
Headless Chrome is a new feature in Chrome that runs it in a headless
environment - no GUI, but with all the features of Chrome through the
CLI!

The purpose of a headless browser is for automated testing and server
environments where you don't need a visible UI shell. For example, you
may want to run some tests against a real web page, create a PDF of it,
or just inspect how the browser renders a URL. There are many options,
consult the docs for more info.

Verified to work on Mac OSX, have not tested for Windows.
Windows release was supposed to come with Chrome 60, however I’ve heard
there may still be issues.